move static status to LicenseUtils
